+++ bookmem/metadata_enrichment.py
@@ -23,6 +23,7 @@ import requests
 
 from .frontmatter import read_markdown_with_frontmatter, write_markdown_with_frontmatter
 from .loc import lookup_loc_by_isbn, enrich_file_with_loc
+from .audit import append_audit_record
 
 
 ENRICHMENT_VERSION = "0.1.0"